requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3

loguru~=0.6.0
selenium~=4.1.3
opencv-python~=4.5.5.64
beautifulsoup4~=4.10.0
qrtools~=0.0.2
Pillow~=9.1.0
pyzbar~=0.1.8